app/Http/Requests/Instructor/Notificationディレクトリが肥大化しているための整理(yuki)
