+++ app/Http/Requests/Instructor/Notification/BulkDeleteRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests\Instructor;
+namespace App\Http\Requests\Instructor\Notification;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class NotificationBulkDeleteRequest extends FormRequest
+class BulkDeleteRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
