+++ app/Http/Requests/Instructor/Notification/IndexRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests\Instructor;
+namespace App\Http\Requests\Instructor\Notification;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class NotificationDeleteRequest extends FormRequest
+class IndexRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -16,13 +16,6 @@ class NotificationDeleteRequest extends FormRequest
         return true;
     }
 
-    protected function prepareForValidation()
-    {
-        $this->merge([
-            'notification_id' => $this->route('notification_id'),
-        ]);
-    }
-
     /**
      * Get the validation rules that apply to the request.
      *
@@ -31,7 +24,8 @@ class NotificationDeleteRequest extends FormRequest
     public function rules()
     {
         return [
-            'notification_id' => ['required', 'integer', 'exists:notifications,id,deleted_at,NULL'],
+            'per_page' => ['integer', 'min:1'],
+            'page' => ['integer', 'min:1'],
         ];
     }
 }
